requirements.txt
tensorflow==2.11.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.3
seaborn==0.12.2
scikit-learn==1.2.2
jupyter==1.0.0
flask
opencv-python
pillow
imbalanced-learn
